requirements.txt
flask==2.3.2
python-dotenv==1.0.0
pytest==7.4.2
gunicorn
bcr_api
sqlalchemy>=2.0
sqlalchemy-bigquery>=1.10.0
google-cloud-bigquery>=3.25.0
pydata-google-auth>=1.8.2

